Kia EU: scaffold KiaEuropeAPIClient file layout
Splits the single-file stub into the same four-file layout used by
HyundaiEuropeAPIClient (main + Headers + Commands + Parsing) and seeds
the main class with the constants from KiaUvoApiEU (PR #1123 in
hyundai_kia_connect_api). All methods still throw regionNotSupported;
the implementation lands in subsequent commits.
