+++ Sources/BetterBlueKit/API/KiaEurope/KiaEuropeAPIClient+Commands.swift
@@ -1,0 +1,13 @@
+//
+//  KiaEuropeAPIClient+Commands.swift
+//  BetterBlueKit
+//
+//  Command helpers for the Kia Europe client.
+//  Filled in during Fase 4.
+//
+
+import Foundation
+
+extension KiaEuropeAPIClient {
+    // commandPathAndBody(for:ccs2:) lands here in Fase 4.
+}
