+++ Sources/BetterBlueKit/API/KiaEurope/KiaEuropeAPIClient+Headers.swift
@@ -1,0 +1,14 @@
+//
+//  KiaEuropeAPIClient+Headers.swift
+//  BetterBlueKit
+//
+//  Header builders + stamp signature helper for the Kia Europe client.
+//  Filled in during Fase 2/3.
+//
+
+import CryptoKit
+import Foundation
+
+extension KiaEuropeAPIClient {
+    // Header builders land here in Fase 2 (login flow) and Fase 3 (authorized requests).
+}
