+++ Sources/BetterBlueKit/API/KiaEurope/KiaEuropeAPIClient+Parsing.swift
@@ -1,0 +1,13 @@
+//
+//  KiaEuropeAPIClient+Parsing.swift
+//  BetterBlueKit
+//
+//  Response parsing for the Kia Europe client.
+//  Filled in during Fase 2/3.
+//
+
+import Foundation
+
+extension KiaEuropeAPIClient {
+    // parseAuthToken / parseVehiclesResponse / parseVehicleStatusResponse land here in Fase 2/3.
+}
